refactor frontend to no longer need to reload page to see command changes
